# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts API",
    "description": "contacts",
    "version": "1.0.0"
  },
  "host": "https://contacts-icp7.onrender.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/contacts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a new contact",
        "parameters":[
          {
            "name":"body",
            "in":"body",
            "required":true,
            "schema":{
              "type":"object",
              "properties":{
                "firstName":{
                  "type":"string"},
                "lastName":{
                  "type":"string"
                },
                "email":{
                  "type":"string"
                },
                "favoriteColor":{
                  "type":"string"
                },
                "birthday":{
                  "type":"string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
"put": {
    "summary": "Update a contact",
    "description": "Update an existing contact by ID",
    "parameters": [
        {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
                "type": "string"
            },
            "description": "The ID of the contact to update"
        },
        {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "favoriteColor":{
                      "type":"string"
                    },
                    "birthday":{
                      "type":"string"
                    }
                }
            },
            "description": "The updated contact information"
        }
    ],
    "responses": {
        "200": {
            "description": "Successful response",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "OK"
                    }
                }
            }
        },
        "400": {
            "description": "Bad Request",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "Bad Request"
                    }
                }
            }
        },
        "404": {
            "description": "Not Found",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "Not Found"
                    }
                }
            }
        },
        "500": {
            "description": "Internal Server Error",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "Internal Server Error"
                    }
                }
            }
        }
    }
}
,
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}